fix bug in else-free ifs

+++ src/interp/interp.h
@@ -154,9 +154,9 @@ struct web49_read_block_state_t {
     web49_interp_t *interp;
     web49_interp_build_t build;
     web49_interp_link_t *links;
-    uint64_t *stack;
-    uint64_t depth;
-    uint64_t nlocals;
+    uint32_t *stack;
+    uint32_t depth;
+    uint32_t nlocals;
     uint64_t alloc_links;
     uint64_t nlinks;
 };
